Fixed Randomness Crashes on 64-bit & Fixed Cached Images

+++ Sexy.TodLib/TodFoley.cpp
@@ -253,7 +253,7 @@ void TodFoley::PlayFoleyPitch(FoleyType theFoleyType, float thePitch)
 		return;
 
 	int aVariations = 0;
-	intptr_t aVariationsArray[10];
+	int aVariationsArray[10];
 	FoleyTypeData* aFoleyData = &mFoleyTypeData[(int)theFoleyType];
 	for (int i = 0; i < 10; i++)
 	{
@@ -266,7 +266,7 @@ void TodFoley::PlayFoleyPitch(FoleyType theFoleyType, float thePitch)
 		}
 	}
 	TOD_ASSERT(aVariations > 0);
-	intptr_t aVariation = TodPickFromArray(aVariationsArray, aVariations);
+	int aVariation = TodPickFromArray(aVariationsArray, aVariations);
 	aFoleyData->mLastVariationPlayed = aVariation;
 	SoundInstance* aSoundInstance = gSexyAppBase->mSoundManager->GetSoundInstance(*aFoleyParams->mSfxID[aVariation]);
 	if (aSoundInstance == nullptr)
